requirements.txt
python-telegram-bot==20.3
ccxt==3.0.71